add some useless code

+++ net/utils.c
@@ -1,11 +1,11 @@
 #include "utils.h"
 
-uint8_t checksum(const unsigned char* addr, int len) {
+uint16_t checksum(const unsigned char* addr, int len) {
 
     int nleft = len;
-    const uint8_t *w = (const uint8_t*) addr;
-    uint16_t sum = 0;
-    uint8_t answer = 0;
+    const uint16_t *w = (const uint8_t*) addr;
+    uint32_t sum = 0;
+    uint16_t answer = 0;
 
     while(nleft > 1) {
         sum += *w++;
